requirements.txt
apprise==1.8.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
gevent==24.2.1
greenlet==3.0.3
idna==3.7
load-dotenv==0.1.0
Markdown==3.6
oauthlib==3.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.1.1
urllib3==2.2.2
websocket==0.2.1
zope.event==5.0
zope.interface==6.4.post2
